Add function roundToNearest
Add a function roundToNearest which rounds num1 to the nearest num2. For example, if you do (round 90 to the nearest 100) it will return 100. This would be a very useful function for creators because instead of using (calculate (floor of (calculate(calculate(x / y) + 0.5) ) * y)) they can just do (round x to the nearest y).